Fix first-degree-hashing and n-quad serialization
The newline at the end of an n-quad is part of the serialization of a quad, it shouldn't
be added in at the end. Also, I wasn't properly transforming the quads for hashing
because I was looking up the :type of a keyword instead of the :type of the quad component.