readable test output and cleanup
Swapped the order of the test clauses so the failure reports make more sense. Also some
whitespace cleanup.

Also changed the signature of `parse-map` to match all of the other "parse-" type
functions, accepting the "externals" arg as the second arg instead of first. Also fixed
up some cases where weren't passing on the base context, and gave it a clearer name.